# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API Auth / UserManager",
    "version": "1.0.0",
    "description": "Documentation de l'API Auth / UserManager (SIWE, OTP, JWT, KYC)"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Serveur local"
    }
  ],
  "paths": {
...
    "/api/users/wallet/{address}/login-status": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Vérifier le statut de connexion d'un utilisateur",
        "description": "Vérifie si un utilisateur est connecté et récupère les informations de sa session active",
        "operationId": "checkLoginStatus",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "required": true,
            "description": "L'adresse Ethereum du wallet",
            "schema": {
              "type": "string",
              "example": "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Statut récupéré avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "oneOf": [
                        {
                          "type": "object",
                          "properties": {
                            "isLoggedIn": {
                              "type": "boolean",
                              "example": true
                            },
                            "session": {
                              "type": "object",
                              "properties": {
                                "email": {
                                  "type": "string"
                                },
                                "loginTime": {
                                  "type": "string"
                                },
                                "lastActivity": {
                                  "type": "string"
                                }
                              }
                            }
                          }
                        },
                        {
                          "type": "object",
                          "properties": {
                            "isLoggedIn": {
                              "type": "boolean",
                              "example": false
                            }
                          }
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Format d'adresse invalide"
          },
          "500": {
            "description": "Erreur serveur"
          }
        }
      }
    },
    "/api/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Récupérer tous les utilisateurs avec pagination",
        "description": "Récupère une liste paginée de tous les identifiants utilisateurs (userIds)",
        "operationId": "getAllUsers",
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "description": "Position de départ pour la pagination",
            "schema": {
              "type": "integer",
              "default": 0,
              "example": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Nombre maximum d'éléments à retourner",
            "schema": {
              "type": "integer",
              "default": 100,
              "maximum": 1000,
              "example": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des utilisateurs récupérée avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "string",
                          "example": "150"
                        },
                        "offset": {
                          "type": "integer",
                          "example": 0
                        },
                        "limit": {
                          "type": "integer",
                          "example": 10
                        },
                        "userIds": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          },
                          "example": [
                            "0x1234...abcd",
                            "0x5678...efgh"
                          ]
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erreur serveur"
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Mettre à jour les informations utilisateur",
        "description": "Permet à un utilisateur connecté de mettre à jour son email, pays et métadonnées.\nL'utilisateur doit être connecté pour effectuer cette opération.\n",
        "operationId": "updateUserInfo",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "privateKey",
                  "email",
                  "country"
                ],
                "properties": {
                  "privateKey": {
                    "type": "string",
                    "description": "Clé privée du wallet",
                    "example": "0x1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Nouvel email",
                    "example": "john.updated@example.com"
                  },
                  "country": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 2,
                    "description": "Nouveau code pays (2 lettres)",
                    "example": "US"
                  },
                  "metadata": {
                    "type": "string",
                    "description": "Nouvelles métadonnées JSON",
                    "example": "{\"firstName\":\"John\",\"lastName\":\"Doe Updated\"}"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Informations mises à jour avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "transactionHash": {
                          "type": "string"
                        },
                        "blockNumber": {
                          "type": "number"
                        },
                        "message": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Données invalides ou email déjà utilisé"
          },
          "401": {
            "description": "Non connecté"
          },
          "500": {
            "description": "Erreur blockchain"
          }
        }
      }
    },
...
    "/api/users/id/{userId}/deactivate": {
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Désactiver un utilisateur",
        "description": "Désactive le compte d'un utilisateur. L'utilisateur ne pourra plus se connecter\nni effectuer d'opérations jusqu'à réactivation.\n**Nécessite les droits Admin.**\n",
        "operationId": "deactivateUser",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "description": "L'identifiant unique de l'utilisateur (bytes32)",
            "schema": {
              "type": "string",
              "example": "0x1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Utilisateur désactivé avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "transactionHash": {
                          "type": "string"
                        },
                        "blockNumber": {
                          "type": "number"
                        },
                        "message": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "userId invalide"
          },
          "401": {
            "description": "Non autorisé"
          },
          "404": {
            "description": "Utilisateur non trouvé"
          },
          "500": {
            "description": "Erreur blockchain"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtenu lors du login"
      }
    }
  },
...
}